requirements.txt
absl-py==0.12.0
argon2-cffi==20.1.0
astor==0.8.1
async-generator==1.10
attrs==21.2.0
backcall==0.2.0
bleach==1.5.0
blessings==1.7
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
dataclasses==0.8
decorator==5.0.7
defusedxml==0.7.1
docopt==0.6.2
entrypoints==0.3
faiss-cpu==1.7.0
gast==0.4.0
grpcio==1.37.1
html5lib==0.9999999
idna==2.10
importlib-metadata==4.0.1
ipykernel==5.5.4
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==3.0.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.2.0
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
Markdown==3.3.4
MarkupSafe==2.0.0
mistune==0.8.4
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.3.0
numpy==1.19.5
nvidia-ml-py3==7.352.0
pandas==1.1.5
pandocfilters==1.4.3
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
pipreqs==0.4.10
prometheus-client==0.10.1
prompt-toolkit==3.0.18
protobuf==3.16.0
psutil==5.8.0
ptyprocess==0.7.0
pycparser==2.20
Pygments==2.9.0
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2021.1
pyzmq==22.0.3
qtconsole==5.1.0
QtPy==1.9.0
requests==2.25.1
Send2Trash==1.5.0
six==1.16.0
tensorboard==1.8.0
tensorflow-gpu==1.8.0
termcolor==1.1.0
terminado==0.9.5
testpath==0.4.4
tornado==6.1
traitlets==4.3.3
typing-extensions==3.10.0.0
urllib3==1.26.4
wcwidth==0.2.5
Werkzeug==2.0.0
widgetsnbextension==3.5.1
yarg==0.1.9
zipp==3.4.1
